refactor: move store to core/store
