+++ src/components/Sidebar.tsx
@@ -2,7 +2,7 @@
 import './Sidebar.css'
 import piecesConfig from '../data/pieces-config.json'
 import type { PiecesConfig, PieceCategory } from '../core/types'
-import { useStore } from '../store/useStore'
+import { useStore } from '../core/store/useStore'
 
 const config = piecesConfig as PiecesConfig
