+++ src/core/scene/HitPlane.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import type { ThreeEvent } from '@react-three/fiber'
-import { useStore } from '../../store/useStore'
+import { useStore } from '../store/useStore'
 import { canPlace, canPlaceTriSnap, canPlaceTriSnapEdge, canPlaceSquareSnap, canPlaceSquareSnapEdge } from '../../utils/validation'
 import { detectSide, getCellPieceShape } from './pieceGeometry'
 import piecesConfig from '../../data/pieces-config.json'
