+++ src/core/scene/Viewport.tsx
@@ -7,7 +7,7 @@ import HullMesh from '../../scene/HullMesh'
 import SceneGrid from './SceneGrid'
 import PlacedPieces from './PlacedPieces'
 import HitPlane, { TriHitPlane } from './HitPlane'
-import { useStore } from '../../store/useStore'
+import { useStore } from '../store/useStore'
 import piecesConfig from '../../data/pieces-config.json'
 import type { PiecesConfig } from '../types'
